Agregando archivos necesarios para adaptar coleccion 'Users'.

+++ src/services/productService.js
@@ -11,9 +11,23 @@ const productService = {
     return await product.save();
   },
   updateProduct: async ({ _id, ...rest }) => {
+    const product = await Product.findById(_id);
+    if (!product) throw new Error('Product not found');
+    
+    if (product.facturapiid){
+      await facturapi.updateProduct(product.facturapiid, {
+        description: rest.description || product.description,
+        price: rest.price || product.price
+      });
+    }
+
     return await Product.findByIdAndUpdate(_id, rest, { new: true });
+
   },
   deleteProduct: async (_id) => {
+    const product = await Product.findById(_id);
+    if (!product) throw new Error('Product not found');
+    if (product.facturapiid){ await facturapi.deleteProduct(product.facturapiid); }
     return await Product.findByIdAndDelete(_id);
   }
 };
